User movement of scrubber in SeekBar calls seekTo on MediaSession.

+++ app/src/main/java/com/example/android/mediasession/service/PlayerAdapter.java
@@ -40,5 +40,5 @@ public interface PlayerAdapter {
 
     void initializeProgressCallback();
 
-    void seekTo(int position);
+    void seekTo(long position);
 }
